Add files via upload
The histogram of Ticket Priority offers details on how the customer support tickets dataset's ticket priorities are distributed. 
X-axis depicts the incompatible ticket priorities. Each bar corresponds to a specific ticket priority category and Y-axis depicts the count or frequency of tickets for each ticket priority category. It shows the number of tickets falling into each priority category.
We can notice which ticket priority categories have an elevated count or frequency in customer support tickets. This visualization assists in recognizing any imbalances or patterns in ticket priorities, which can be valuable for future surveys and arriving at conclusions in customer support operations.

The scatter plot of ticket priority and ticket channel gives information about the distribution, patterns, and relationships between the data points. The position of the dot accord with the values of ticket priority and ticket channel for that particular data point.
X-axis depicts the ticket priority, specifying the priority level assigned to each customer support ticket and Y-axis depicts the ticket channel, indicating the communication channel through which the ticket was received.
We can notice if determined ticket channels are related to higher or lower priority levels and this information assists in recognizing any trends or patterns in how tickets are organized based on the channel of communication.